refactor: 더미데이터와 컴포넌트 로직을 분리하여 구조 개선
