+++ src/components/modal/schedule_modal/ScheduleCalendar.tsx
@@ -1,0 +1,50 @@
+/** @jsxImportSource @emotion/react */
+import { css } from "@emotion/react";
+import { DayPicker } from "react-day-picker";
+import "react-day-picker/dist/style.css";
+import { useThemeColors } from "../../../hooks/useThemeColors";
+
+interface ScheduleCalendarProps {
+  selectedDate: Date | undefined;
+  onSelect: (date: Date | undefined) => void;
+}
+
+export const ScheduleCalendar = ({
+  selectedDate,
+  onSelect,
+} : ScheduleCalendarProps) => {
+  const { headerBorder } = useThemeColors();
+
+  const calendarWrapper = css`
+    flex: 0 0 280px;
+    border-right: 1px solid ${headerBorder};
+    padding: 16px;
+    height: 100%;
+    box-sizing: border-box;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  `;
+
+  const calendarBox = css`
+    width: 100%;
+    height: 100%;
+  `;
+
+  return (
+    <div css={calendarWrapper}>
+      <div css={calendarBox}>
+        <DayPicker
+          mode="single"
+          selected={selectedDate}
+          onSelect={onSelect}
+          styles={{
+            root: { height: "100%" },
+          }}
+        />
+      </div>
+    </div>
+  );
+};
+
+export default ScheduleCalendar;
