+++ src/components/modal/schedule_modal/scheduleModal.tsx
@@ -1,28 +1,23 @@
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
 import React, { useState, useEffect } from "react";
-import { IoIosClose } from "react-icons/io";
-import { MdCalendarToday } from "react-icons/md";
-import { DayPicker } from "react-day-picker";
-import "react-day-picker/dist/style.css";
 import { ScheduleList } from "./ScheduleList";
 import { ScheduleForm } from "./ScheduleForm";
 import type { Schedule, ScheduleFormData, ViewType } from "./types/schedule";
 import { scheduleAPI } from "../../../api/schedule";
 import { useThemeColors } from "../../../hooks/useThemeColors";
+import ScheduleHeader from "./ScheduleHeader";
+import ScheduleCalendar from "./ScheduleCalendar";
 
-const ScheduleModal: React.FC = () => {
-const [view, setView] = useState<ViewType>("list"); // 현재 화면 (list or form)
-const [schedules, setSchedules] = useState<Schedule[]>([]); // 일정 목록
-const [selectedDate, setSelectedDate] = useState<Date | undefined>(new Date()); // 선택한 날짜
-const [editingSchedule, setEditingSchedule] = useState<Schedule | null>(null); // 수정 중인 일정
-const [loading, setLoading] = useState<boolean>(false); // 로딩 상태
+const ScheduleModal = () => {
+  const [view, setView] = useState<ViewType>("list");
+  const [schedules, setSchedules] = useState<Schedule[]>([]);
+  const [selectedDate, setSelectedDate] = useState<Date | undefined>(new Date());
+  const [editingSchedule, setEditingSchedule] = useState<Schedule | null>(null);
+  const [loading, setLoading] = useState<boolean>(false);
 
+  const { modalBackground } = useThemeColors();
 
-  const { modalBackground, headerBorder, modalHeaderBg, descriptionText } =
-    useThemeColors();
-
-  // 스타일 정의
   const modalContainer = css`
     background: ${modalBackground};
     border-radius: 16px;
@@ -36,52 +31,10 @@ const [loading, setLoading] = useState<boolean>(false); // 로딩 상태
     flex-direction: column;
   `;
 
-  const header = css`
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 16px 24px;
-    border-bottom: 1px solid ${headerBorder};
-    background: ${modalHeaderBg};
-  `;
-
-  const headerTitle = css`
-    display: flex;
-    align-items: center;
-    gap: 12px;
-
-    h2 {
-      font-size: 20px;
-      font-weight: 600;
-      margin: 0;
-    }
-
-    span {
-      color: ${descriptionText};
-      font-size: 14px;
-    }
-  `;
-
   const body = css`
     display: flex;
-    flex: 1; 
+    flex: 1;
     overflow: hidden;
-  `;
-
-  const calendarWrapper = css`
-    flex: 0 0 280px; 
-    border-right: 1px solid ${headerBorder};
-    padding: 16px;
-    height: 100%; 
-    box-sizing: border-box;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  `;
-
-  const calendarBox = css`
-    width: 100%;
-    height: 100%;
   `;
 
   const content = css`
@@ -90,17 +43,13 @@ const [loading, setLoading] = useState<boolean>(false); // 로딩 상태
     overflow-y: auto;
   `;
 
-  const clickButton = css`
-    cursor: pointer;
-  `;
-
   // 일정 목록 로드
   const loadSchedules = async (): Promise<void> => {
     if (!selectedDate) return;
     setLoading(true);
     try {
-      const dateStr = selectedDate.toLocaleDateString("sv-SE"); // YYYY-MM-DD 형식
-      const data = await scheduleAPI.getSchedules(dateStr); // 일정 불러오기
+      const dateStr = selectedDate.toLocaleDateString("sv-SE"); // YYYY-MM-DD
+      const data = await scheduleAPI.getSchedules(dateStr);
       setSchedules(data);
     } catch (error) {
       console.error("일정 로드 실패:", error);
@@ -109,7 +58,6 @@ const [loading, setLoading] = useState<boolean>(false); // 로딩 상태
     }
   };
 
-  // 선택한 날짜가 바뀌면 일정 다시 로드
   useEffect(() => {
     loadSchedules();
   }, [selectedDate]);
@@ -165,13 +113,11 @@ const [loading, setLoading] = useState<boolean>(false); // 로딩 상태
           formData
         );
         if (result.success && result.data) {
-          // 업데이트 후 목록 다시 불러오기
           await loadSchedules();
         }
       } else {
         const result = await scheduleAPI.createSchedule(formData);
         if (result.success && result.data) {
-          // 새 일정 생성 후 목록 다시 불러오기
           await loadSchedules();
         }
       }
@@ -192,30 +138,13 @@ const [loading, setLoading] = useState<boolean>(false); // 로딩 상태
 
   return (
     <div css={modalContainer}>
-      <div css={header}>
-        <div css={headerTitle}>
-          <MdCalendarToday />
-          <div>
-            <h2>일정관리</h2>
-            <span>스케줄을 효율적으로 관리하세요</span>
-          </div>
-        </div>
-        <IoIosClose size={20} css={clickButton} />
-      </div>
+      <ScheduleHeader />
 
       <div css={body}>
-        <div css={calendarWrapper}>
-          <div css={calendarBox}>
-            <DayPicker
-              mode="single"
-              selected={selectedDate}
-              onSelect={setSelectedDate}
-              styles={{
-                root: { height: "100%" },
-              }}
-            />
-          </div>
-        </div>
+        <ScheduleCalendar
+          selectedDate={selectedDate}
+          onSelect={setSelectedDate}
+        />
 
         <div css={content}>
           {view === "list" ? (
